backend/requirements-basic.txt
# Basic working requirements for deployment
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.42
python-dotenv==1.0.1
supabase==2.11.3

# LINE Bot
line-bot-sdk==3.17.1

# Google AI & Langchain
google-generativeai==0.8.3
langchain==0.3.18
langchain-google-genai==2.0.17
langchain-core==0.3.35
langgraph==0.2.111

# Basic dependencies
pydantic==2.11.7
requests==2.32.4
aiofiles==23.2.1
gunicorn==21.2.0
langgraph==0.0.62 
langchain==0.1.20 
langchain-google-genai==0.0.11 
langchain-google-genai==0.0.11 
langchain-google-genai==0.0.11 
langchain-core==0.1.52 
